add visualcrossing forecast source
use a common cache directory for all sources
all sources in same package for now
visualcrossing provides a 15 day forecast, with 3 hour resolution after 7 days.
because of that, we avoid writing null data points (if temp is null in json)
this is also why we have pointers for most values in the parsed struct.
snow is a special case that is set to nil when there is no snow coming sometimes.
we have to calculate dewpoint from rh and t since visualcrossing does not provide it.
we have to create feelslike from temp/windchill/heatindex because there is no combined version
created an app type to hold program fields, reducing parameters in main package