Remove proxy ability.
Instead of using the proxy, I'm using a configurable datasource in grafana.
Currently the queries I make to ForecastMetrics and VictoriaMetrics are not resolvable by each other, so by switching the datasource, I'm switching which queries resolve.
This is much less complicated than trying to handle special cases when proxying.
Also removes a bunch of todo items.